requirements.txt
jsons==1.4.2
keyboard==0.13.5
pretty-downloader==0.0.9
PySimpleGUI==4.46.0
pystray==0.17.3
Pillow==8.3.1
pywin32==301
PyYAML==5.4.1
watchdog==2.1.3
inject==4.3.1
winregal==1.0.1
hurry.filesize==0.9
natsort==7.1.1
